pyproject.toml
[project]
name = "habitica"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "eventlet (>=0.40.4,<0.41.0)",
    "drf-yasg (>=1.21.11,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0) ; python_version >= \"3.13\" and python_version < \"4.0\"",
    "coverage (>=7.12.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.11.0"
isort = "^7.0.0"

